+++ lms_android/app/src/main/java/hse/diploma/cybersecplatform/ui/screens/tasks/TasksScreen.kt
@@ -1,4 +1,4 @@
-package hse.diploma.cybersecplatform.ui.screens.home
+package hse.diploma.cybersecplatform.ui.screens.tasks
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Spacer
